Avoid baud truncation.

+++ vehicle_interface_controller/common/usart.hpp
@@ -7,7 +7,7 @@
 namespace usart {
     constexpr uint32_t clockFrequency = F_CPU;
 
-    void beginAtBaud(uint16_t baud) {
+    void beginAtBaud(uint32_t baud) {
         uint8_t ubrr = clockFrequency/16/baud - 1;
         UBRR0H = ubrr >> 8;
         UBRR0L = ubrr;
